added initial commit for spawning tmux profiles
I need to add this to my gh, I want to have it reproduceable, but
failing still, because it's just copy pasting the profile, I'm thinking
of using the configuration as is, but change the gen tooling for firefox to a nixos
container.

This is what I currently use.. So PR's on this are welcome, but will not
be committed fast probably. ;)