Updated README.md, Started ULA usage in opnsense
REAMDE.md:
Big update on the Readme, the plan is to make this the standard where i
can fall back on.

Nixos config:
Not a fully working setup, but i get the prefix from opnsense now
without dhcpv6, still need to make some routing in nixos.

The tools-debugging will be a general nix file to include, not really
sure how i will structure this.